refactor(backend): use uint64 for count fields and improve null handling

+++ internal/infrastructure/repository/observability/score_repository.go
@@ -399,7 +399,7 @@ func (r *scoreRepository) GetAggregationsByExperiments(
 			StdDev: stddevValue,
 			Min:    minValue,
 			Max:    maxValue,
-			Count:  int64(count),
+			Count:  count,
 		}
 	}
